requirements.txt
yt-dlp==2024.3.10
openai-whisper==20231117
deep-translator==1.11.4